requirements.txt
Jinja2==3.1.4
WeasyPrint==62.3
openai==1.68.2
python-dotenv==1.0.1
